+++ dflib/src/main/java/com/nhl/dflib/exp/num/DecimalExp1.java
@@ -13,6 +13,13 @@ import java.util.function.Function;
  */
 public class DecimalExp1<F> extends MapExp1<F, BigDecimal> implements DecimalExp {
 
+    /**
+     * @since 0.14
+     */
+    public static <F> DecimalExp1<F> map(String opName, Exp<F> exp, Function<Series<F>, Series<BigDecimal>> op) {
+        return new DecimalExp1<>(opName, exp, op);
+    }
+
     public static <F> DecimalExp1<F> mapVal(String opName, Exp<F> exp, Function<F, BigDecimal> op) {
         return new DecimalExp1<>(opName, exp, valToSeries(op));
     }