+++ dflib/src/main/java/com/nhl/dflib/exp/num/DoubleExpFactory.java
@@ -1,6 +1,10 @@
 package com.nhl.dflib.exp.num;
 
-import com.nhl.dflib.*;
+import com.nhl.dflib.Condition;
+import com.nhl.dflib.DecimalExp;
+import com.nhl.dflib.DoubleSeries;
+import com.nhl.dflib.Exp;
+import com.nhl.dflib.NumExp;
 import com.nhl.dflib.exp.agg.DoubleAggregators;
 import com.nhl.dflib.exp.agg.DoubleExpAggregator;
 
@@ -86,6 +90,11 @@ public class DoubleExpFactory extends NumericExpFactory {
     @Override
     public DecimalExp castAsDecimal(NumExp<?> exp) {
         return DecimalExp1.mapVal("castAsDecimal", cast(exp), BigDecimal::valueOf);
+    }
+
+    @Override
+    public NumExp<Double> cumSum(Exp<? extends Number> exp) {
+        return DoubleExp1.map("cumSum", exp, DoubleAggregators::cumSum);
     }
 
     @Override